fix: add configuration module to configure java sender

+++ spring/src/main/java/com/unilocal/backend/service/PlaceService.java
@@ -3,7 +3,6 @@ package com.unilocal.backend.service;
 import com.unilocal.backend.models.Place;
 import com.unilocal.backend.repos.PlaceRepository;
 import java.util.List;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.annotation.Id;
 import org.springframework.stereotype.Service;
@@ -11,27 +10,22 @@ import org.springframework.stereotype.Service;
 @Service
 public class PlaceService {
 
-  @Autowired
-  PlaceRepository placeRepository;
+  @Autowired PlaceRepository placeRepository;
 
   // return a list of all places
-  public List<Place> getAll() {
-    return placeRepository.findAll();
-  }
-
-  // save a new place
-  public Place save(Place place) {
-    return placeRepository.save(place);
-  }
+  public List<Place> getAll() { return placeRepository.findAll(); }
 
   // find a place by its id
-  public Place findById(Id id) {
-    return placeRepository.findById(id);
-  }
+  public Place findById(Id id) { return placeRepository.findById(id); }
+
+  // save a new place
+  public Place save(Place place) { return placeRepository.save(place); }
 
   // delete a place by its id
   public void delete(Id id) {
     Place place = placeRepository.findById(id);
+    if (place == null)
+      throw new RuntimeException("Place not found");
     placeRepository.delete(place);
   }
 }
